3.3-permissions/api_with_restrictions/requirements.txt
Django==4.2.3
django-filter==23.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5